*) fixed base data type definition.

+++ src/os/win32/ngx_recv.c
@@ -43,7 +43,7 @@ retry:
         ovlp = NULL;
 
         wsabuf.buf = buf;
-        wsabuf.len = size;
+        wsabuf.len = (ULONG) size;
     }
 
     n = 0;
